backend/auth/requirements.txt
nameko==3.0.0-rc11
python-jose==3.1.0
python-dotenv
pytest
pytest-cov
sqlmodel
sqlmodel-controller
redis
cryptography
jose
bcrypt